+++ src/resources/boards/board.router.js
@@ -1,8 +1,8 @@
-import { readController } from './controllers/read.js';
-import { readByIdController } from './controllers/read-by-id.js';
-import { createController } from './controllers/create.js';
-import { deleteByIdController } from './controllers/delete-by-id.js';
-import { updateByIdController } from './controllers/update-by-id.js';
+import { readController } from './controllers/board.read.js';
+import { readByIdController } from './controllers/board.read-by-id.js';
+import { createController } from './controllers/board.create.js';
+import { deleteByIdController } from './controllers/board.delete-by-id.js';
+import { updateByIdController } from './controllers/board.update-by-id.js';
 
 const BOARD_ID = 'boardId';
