+++ src/resources/tasks/controllers/task.delete-by-id.js
