+++ src/resources/users/user.router.js
@@ -1,8 +1,8 @@
-import { readController } from './controllers/read.js';
-import { readByIdController } from './controllers/read-by-id.js';
-import { createController } from './controllers/create.js';
-import { deleteByIdController } from './controllers/delete-by-id.js';
-import { updateByIdController } from './controllers/update-by-id.js';
+import { readController } from './controllers/user.read.js';
+import { readByIdController } from './controllers/user.read-by-id.js';
+import { createController } from './controllers/user.create.js';
+import { deleteByIdController } from './controllers/user.delete-by-id.js';
+import { updateByIdController } from './controllers/user.update-by-id.js';
 
 const USER_ID = 'userId';
